Fix line mapping being incorrect.
Until now, a wrong line mapping would be computed when adding text e.g. to the
end of a line.

This is because in the `git diff --word-diff=porcelain` that was used so far,
such changes appear as a "+" line (which makes sense given that it's a word-
diff), which the code so far interpreted as the addition of a new line when in
fact only words inside the line were changed.

So far, I believe that a correct line mapping cannot be computed using
`--word-diff`, because it inherently ignores whitespace: If you delete
whitespace at the end of the line, and also some following empty lines,
it doesn't even appear in `git diff --word-diff=porcelain`.

This commit uses `diff` instead, which has a feature to tell us the
added, deleted and unchanged lines, from which we can directly
construct the line mapping.